requirements.txt
# Convexity AI - Dependencies

# UI
streamlit>=1.30.0
plotly>=5.18.0

# Data
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0

# Database
lancedb>=0.4.0
# sqlite3 is built into Python

# AI
groq>=0.4.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0  # For Nano Banana image generation
stitch-ai>=0.3.0            # For AI memory management

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule>=1.2.0

# Optional (for email)
# smtplib is built into Python
